+++ src/Sulu/Bundle/ContentBundle/Content/Structure/ExcerptStructureExtension.php
@@ -16,14 +16,18 @@ use Sulu\Bundle\SearchBundle\Search\Factory;
 use Sulu\Component\Content\Compat\PropertyInterface;
 use Sulu\Component\Content\Compat\StructureInterface;
 use Sulu\Component\Content\Compat\StructureManagerInterface;
+use Sulu\Component\Content\ContentTypeExportInterface;
 use Sulu\Component\Content\ContentTypeManagerInterface;
+use Sulu\Component\Content\Export\ContentExportManagerInterface;
 use Sulu\Component\Content\Extension\AbstractExtension;
+use Sulu\Component\Content\Extension\ExportExtensionInterface;
+use Sulu\Component\Content\Import\ContentImportManagerInterface;
 use Sulu\Component\Content\Mapper\Translation\TranslatedProperty;
 
 /**
  * extends structure with seo content.
  */
-class ExcerptStructureExtension extends AbstractExtension
+class ExcerptStructureExtension extends AbstractExtension implements ExportExtensionInterface
 {
     /**
      * name of structure extension.
@@ -57,6 +61,16 @@ class ExcerptStructureExtension extends AbstractExtension
     protected $structureManager;
 
     /**
+     * @var ContentExportManagerInterface
+     */
+    protected $contentExportManager;
+
+    /**
+     * @var ContentImportManager
+     */
+    protected $contentImportManager;
+
+    /**
      * @var string
      */
     private $languageNamespace;
@@ -74,11 +88,15 @@ class ExcerptStructureExtension extends AbstractExtension
     public function __construct(
         StructureManagerInterface $structureManager,
         ContentTypeManagerInterface $contentTypeManager,
+        ContentExportManagerInterface $contentExportManager,
+        ContentImportManagerInterface $contentImportManager,
         Factory $factory
     ) {
         $this->contentTypeManager = $contentTypeManager;
         $this->structureManager = $structureManager;
+        $this->contentImportManager = $contentImportManager;
         $this->factory = $factory;
+        $this->contentExportManager = $contentExportManager;
     }
 
     /**
@@ -225,6 +243,81 @@ class ExcerptStructureExtension extends AbstractExtension
         /** @var PropertyInterface $property */
         foreach ($this->getExcerptStructure($locale)->getProperties() as $property) {
             $this->properties[] = $property->getName();
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function export($properties, $format = null)
+    {
+        $container = new ExcerptValueContainer($properties);
+
+        $data = [];
+        foreach ($this->getExcerptStructure()->getProperties() as $property) {
+            if ($container->__isset($property->getName())) {
+                $property->setValue($container->__get($property->getName()));
+                $contentType = $this->contentTypeManager->get($property->getContentTypeName());
+                if ($this->contentExportManager->hasExport($property->getContentTypeName(), $format)) {
+                    $options = $this->contentExportManager->getOptions($property->getContentTypeName(), $format);
+
+                    $data[$property->getName()] = [
+                        'name' => $property->getName(),
+                        'value' => $contentType->exportData($property->getValue()),
+                        'type' => $property->getContentTypeName(),
+                        'options' => $options,
+                    ];
+                }
+            }
+        }
+
+        return $data;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getImportPropertyNames()
+    {
+        $propertyNames = [];
+
+        foreach ($this->getExcerptStructure()->getProperties() as $property) {
+            $propertyNames[] = $property->getName();
+        }
+
+        return $propertyNames;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function import(NodeInterface $node, $data, $webspaceKey, $languageCode, $format)
+    {
+        $this->setLanguageCode($languageCode, 'i18n', null);
+
+        foreach ($this->getExcerptStructure()->getProperties() as $property) {
+            $contentType = $this->contentTypeManager->get($property->getContentTypeName());
+
+            if (
+                isset($data[$property->getName()])
+                && $this->contentImportManager->hasImport($property->getContentTypeName(), $format)
+            ) {
+                $property->setValue($data[$property->getName()]);
+                /** @var ContentTypeExportInterface $contentType */
+                $contentType->importData(
+                    $node,
+                    new TranslatedProperty(
+                        $property,
+                        $languageCode,
+                        $this->languageNamespace,
+                        $this->additionalPrefix
+                    ),
+                    null, // userid
+                    $webspaceKey,
+                    $languageCode,
+                    null // segmentkey
+                );
+            }
         }
     }
 }