+++ src/Sulu/Bundle/ContentBundle/DependencyInjection/SuluContentExtension.php
@@ -145,6 +145,7 @@ class SuluContentExtension extends Extension implements PrependExtensionInterfac
         $loader->load('compat.xml');
         $loader->load('document.xml');
         $loader->load('serializer.xml');
+        $loader->load('export.xml');
     }
 
     private function processTemplates(ContainerBuilder $container, $config)