+++ src/Sulu/Bundle/ContentBundle/SuluContentBundle.php
@@ -11,6 +11,7 @@
 
 namespace Sulu\Bundle\ContentBundle;
 
+use Sulu\Bundle\ContentBundle\DependencyInjection\Compiler\ContentExportCompilerPass;
 use Sulu\Bundle\ContentBundle\DependencyInjection\Compiler\SecurityPass;
 use Sulu\Bundle\ContentBundle\DependencyInjection\Compiler\SmartContentDataProviderCompilerPass;
 use Sulu\Bundle\ContentBundle\DependencyInjection\Compiler\StructureExtensionCompilerPass;
@@ -25,6 +26,7 @@ class SuluContentBundle extends Bundle
     {
         parent::build($container);
 
+        $container->addCompilerPass(new ContentExportCompilerPass());
         $container->addCompilerPass(new SecurityPass());
         $container->addCompilerPass(new SmartContentDataProviderCompilerPass());
         $container->addCompilerPass(new WebspacesPass());