+++ src/Sulu/Bundle/SnippetBundle/Content/SnippetContent.php
@@ -21,12 +21,13 @@ use Sulu\Component\Content\Compat\PropertyParameter;
 use Sulu\Component\Content\Compat\Structure\PageBridge;
 use Sulu\Component\Content\Compat\Structure\SnippetBridge;
 use Sulu\Component\Content\ComplexContentType;
+use Sulu\Component\Content\ContentTypeExportInterface;
 use Sulu\Component\Content\ContentTypeInterface;
 
 /**
  * ContentType for Snippets.
  */
-class SnippetContent extends ComplexContentType
+class SnippetContent extends ComplexContentType implements ContentTypeExportInterface
 {
     /**
      * @var SnippetResolverInterface
@@ -100,9 +101,9 @@ class SnippetContent extends ComplexContentType
         $languageCode,
         $segmentKey
     ) {
-        $snippetReferences = [];
         $values = $property->getValue();
 
+        $snippetReferences = [];
         $values = is_array($values) ? $values : [];
 
         foreach ($values as $value) {
@@ -236,5 +237,34 @@ class SnippetContent extends ComplexContentType
         }
 
         return $parameter[$name]->getValue();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function exportData($propertyValue)
+    {
+        $uuids = $this->getUuids($propertyValue);
+
+        if (!empty($uuids)) {
+            return json_encode($this->getUuids($propertyValue));
+        }
+
+        return '';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function importData(
+        NodeInterface $node,
+        PropertyInterface $property,
+        $userId,
+        $webspaceKey,
+        $languageCode,
+        $segmentKey = null
+    ) {
+        $property->setValue(json_decode($property->getValue()));
+        $this->write($node, $property, $userId, $webspaceKey, $languageCode, $segmentKey);
     }
 }