+++ src/Sulu/Component/Content/Export/ContentExportManagerInterface.php
@@ -1,0 +1,47 @@
+<?php
+
+/*
+ * This file is part of Sulu.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Component\Content\Export;
+
+/**
+ * Interface for Content Export Manager.
+ */
+interface ContentExportManagerInterface
+{
+    /**
+     * @param $contentTypeName
+     * @param $format
+     * @param $options
+     */
+    public function add($contentTypeName, $format, $options);
+
+    /**
+     * @param $contentTypeName
+     * @param $propertyValue
+     */
+    public function export($contentTypeName, $propertyValue);
+
+    /**
+     * @param $contentTypeName
+     * @param $format
+     *
+     * @return bool
+     */
+    public function hasExport($contentTypeName, $format);
+
+    /**
+     * @param $contentTypeName
+     * @param $format
+     *
+     * @return array
+     */
+    public function getOptions($contentTypeName, $format);
+}